requirements.txt
streamlit
numpy==1.23.5
pillow
pandas
scikit-learn==1.2.2
xgboost==2.0.0git add streamlitApp.py voting_model.pkl requirements.txt